plt: add support to read the MAC from fuse and write to NVS
Add a new command "plt set_mac" to allow reading the MAC address from
the fuse ROM and writing to the NVS.

To ease the usage, also allow a real MAC address to be passed.  If a
real MAC is passed, we wouldn't need to know the interface to use, but
it's better to have everything in a single command.

If the MAC needs to be assigned offline (ie. not in a device with a
valid wl12xx interface), the old "set nvs_mac" command still works.

Signed-off-by: Luciano Coelho <coelho@ti.com>