plt: add an option to assign the default MAC address to the NVS
The default MAC address is 00:00:00:00:00:00, which means that the
driver will read the address from fuse.

This is redundant with manually assigning the address to zeros, but it
checks whether the hardware supports reading the MAC from fuse ROM and
fails if not.  This brings more awareness to the fact that the MAC may
be unavailable in the fuse.

Signed-off-by: Luciano Coelho <coelho@ti.com>