wlconf: add conf.h file with current structures
This should not be used in production, it is here just as an example
of the small changes that need to be done in the conf.h files in order
to get it to work with wlconf.

The two main things to keep in mind are: arrays with non-numerical
(ie. defines or enums) sizes are not supported; the priv part needs to
be manually included in the same file.

Signed-off-by: Luciano Coelho <coelho@ti.com>