wlconf: add command to generate the binary struct file
In most cases we don't want to parse the entire source file everytime
we run the command.  So we can pre-parse it and write the results to a
binary file that can be used next times.  Add an option -G to parse
the header file and write the structure data into a binary.

Signed-off-by: Luciano Coelho <coelho@ti.com>