wlconf: add default struct and configuration binaries
It is rather cumbersome for the end-user to parse a source header file
and generate the structure binaries.  Add a default struct.bin file
with the correct structures for the current driver.

Also add a default configuration binary to be used when generating
modified ones, instead of reading it from the driver debugfs entry and
change the default input configuration binary accordingly.

Signed-off-by: Luciano Coelho <coelho@ti.com>