wlconf: parse magic and version defines from the struct header
We need to make sure that the binaries we read and generate are
compatible with the structure defined in the source code.

To do this, we have the magic number and the version number.  This
commit parses the defines that specify these values in the struct
source, writes it to the struct binary and reads them from the struct
binary.

Also update the supplied struct.bin file accordingly.

Signed-off-by: Luciano Coelho <coelho@ti.com>