wlconf: add -X option to ignore checksum from conf binary
The driver doesn't write the correct checksum to the debugfs conf
structure yet, so when we update the structure, we need to manually
add the checksum for the tool to work.  This commit adds a -X option
that causes the tool to ignore the checksum value so the actual
checksum can be set (or any other command can be executed).

Signed-off-by: Luciano Coelho <coelho@ti.com>