wlconf: rename parse-ini to parse-text-conf to avoid confusion
We are required to parse previously specified INI files which have a
different format than the one we have defined.  Those files are
already called INI files, so let's call ours text configuration files.

Rename the option to parse them from --parse-ini to --parse-text-conf
and change the short option from -I to -C.  Add a few test cases for
this feature and rename the example.ini file to example.conf.

Signed-off-by: Luciano Coelho <coelho@ti.com>