wlconf: implement parsing of INI files in TI's existing format
Add a new command to read the contents of an INI file and set the
specified values into the output binary.  A new dictionary.txt file is
used in the translation of flat INI element names to internal
structured element names.

Add unit tests to cover some of the INI file reading cases.  Update
README file and add an example.ini file based on a real-life example.

Signed-off-by: Luciano Coelho <coelho@ti.com>