wlconf: merge parse_ini() and parse_text_conf() into a single function
These two functions were doing basically the same thing, with the
difference that the INI version was doing a translation into the
internal conf format and using a different regex.  To avoid lots of
duplicate code, we can pass the file type and let the function decide
what to do.

Signed-off-by: Luciano Coelho <coelho@ti.com>