calibrator: HP_FEM, add support for optional ini file parameters
In current implementation, parsing of ini files expect a certain amount
of parameters/values to exist for each params section. This prevents
optional params - in other words all params must exist

The new ini files for HP FEM has some optional params - i.e. SRF
params do not exist

In order to support this the parsing now checks if we started a new
params section even if we didn't finish the previous one. This in fact
allows for optional params

A log is now printed out to indicate this situation to the user

Signed-off-by: Yair Shapira <yair.shapira@ti.com>
Signed-off-by: Luciano Coelho <coelho@ti.com>