wlconf: Fix parsing of comments
When parsing conf.h wlconf uses regular expression to find and remove
comments.

There was a bug in the regular expression that caused the Android
version to remove also non-comments lines.

For example:
------------
After start of comment (/*) the following was used to skip non end-of-comment:
([^*]|[^*][^/])*

This is buggy because we want to skip anything  not ([*][/]))*
but (not ([*][/])) doesn't translate to [^*][^/] (de-morgan ;-))
It does translate to
[^*] or ([*] but not [^/])

So the correct skipping expression is:
([^*]|[*][^/])*

Signed-off-by: Yair Shapira <yair.shapira@ti.com>