wlconf: regenerated wl18xx-conf-default.bin
file is regenerated from debugfs conf file and in addition
wlconf -X -o ... is used to fix its checksum.

This provides a temporary solution for a problem that exist when
generating with -D option:  values that are array of structs
(i.e. ac_conf) are not written correctly - only first element is
written and the others contain unknown information.

This is an known issue and limitation with current wlconf
implementation.

Signed-off-by: Yair Shapira <yair.shapira@ti.com>