calibrator: add support for plt fem-detect/chip-awake power modes
fem-detect power mode works on wl6/7 and returns the fem radio type.
This is the basis for calibrator auto-fem logic where the decision
on fem type to use is done in calibrator itself and not the driver.

auto-fem logic will be added later on.

chip-awake mode turns the chip on and perform sdio initialization
and ELP wakeup. This mode is used for RTTT sdio bridging.

Signed-off-by: Yair Shapira <yair.shapira@ti.com>